refactor using mod_interface
