+++ module/rust/math_adapter/rust/impl/math/abs/from.rs
@@ -38,30 +38,8 @@ pub( crate ) mod private
 
 }
 
-/// Protected namespace of the module.
-pub mod protected
+crate::mod_interface!
 {
-}
-
-/// Orphan namespace of the module.
-pub mod orphan
-{
-}
-
-/// Exposed namespace of the module.
-pub mod exposed
-{
-  // use super::internal as i;
-  pub use super::private::From2;
-  pub use super::private::Into2;
-}
-
-pub use exposed::*;
-
-/// Prelude to use essentials: `use my_module::prelude::*`.
-pub mod prelude
-{
-  // use super::internal as i;
-  pub use super::private::From2;
-  pub use super::private::Into2;
+  prelude use From2;
+  prelude use Into2;
 }
