+++ module/rust/math_adapter/rust/impl/math/abs/mod.rs
@@ -1,0 +1,9 @@
+meta_tools::mod_interface!
+{
+  /// Interfaces to either conver or reinterpret nath data structure as analog of math lib of choice..
+  layer as_foreign;
+  /// Local implementation of traits From / Into. Required because of limmitations of Rust.
+  layer from;
+  /// Define several traits like NanLikeInterface.
+  layer nan_like;
+}
