+++ module/rust/math_adapter/rust/impl/math/abs/nan_like.rs
@@ -1,6 +1,7 @@
 /// Internal namespace.
 pub( crate ) mod private
 {
+  use crate::*;
 
   ///
   /// Implement check is it nan and constructor with NAN value.
@@ -62,33 +63,12 @@ pub( crate ) mod private
     };
   }
 
-  crate::for_each_int!( impl_nan_like_for_integer );
-  crate::for_each_float!( impl_nan_like_for_float );
+  for_each_int!( impl_nan_like_for_integer );
+  for_each_float!( impl_nan_like_for_float );
 
 }
 
-/// Protected namespace of the module.
-pub mod protected
+crate::mod_interface!
 {
-}
-
-/// Orphan namespace of the module.
-pub mod orphan
-{
-}
-
-/// Exposed namespace of the module.
-pub mod exposed
-{
-  // use super::internal as i;
-  pub use super::private::NanLikeInterface;
-}
-
-pub use exposed::*;
-
-/// Prelude to use essentials: `use my_module::prelude::*`.
-pub mod prelude
-{
-  // // use super::internal as i;
-  // pub use super::private::NanLikeInterface;
+  prelude use NanLikeInterface;
 }
