--- module/rust/math_adapter/rust/impl/math/as_foreign.rs
@@ -1,6 +1,0 @@
-meta_tools::mod_interface!
-{
-  /// As foreign.
-  layer as_foreign;
-}
-
