--- module/rust/math_adapter/rust/impl/math/from.rs
@@ -1,6 +1,0 @@
-meta_tools::mod_interface!
-{
-  /// From.
-  layer from;
-}
-
