--- module/rust/math_adapter/rust/impl/math/interfaces.rs
@@ -1,5 +1,0 @@
-meta_tools::mod_interface!
-{
-  /// Interfaces.
-  layer interfaces;
-}
