--- module/rust/math_adapter/rust/impl/math/macro_tools.rs
@@ -1,2 +1,0 @@
-
-pub use ::meta_tools::dependency::for_each::*;
