+++ module/rust/math_adapter/rust/impl/math/meta/mod.rs
@@ -1,0 +1,5 @@
+meta_tools::mod_interface!
+{
+  /// For each type of numer.
+  layer for_each;
+}
