+++ module/rust/math_adapter/rust/impl/math/vector.rs
@@ -4,7 +4,7 @@ meta_tools::mod_interface!
   /// Basics.
   layer x2;
 }
-pub use x2::macros;
-#[ allow( unused_imports ) ]
-pub( crate ) use x2::macros::*;
 
+// pub use x2::macros;
+// #[ allow( unused_imports ) ]
+// pub( crate ) use x2::macros::*;
