+++ module/rust/math_adapter/rust/impl/math/vector/x2/impl_deref.rs
@@ -1,40 +1,57 @@
-///
-/// Implement dereferencing of regular math object to math objects of math lib of choice.
-///
-
-#[ allow( unused_macros ) ]
-macro_rules! impl_x2_deref
+/// Internal namespace.
+pub( crate ) mod private
 {
+  use crate::*;
 
-  () => {};
 
-  ( $Va : ident $( :: $Vb : ident )* ) =>
+  ///
+  /// Implement dereferencing of regular math object to math objects of math lib of choice.
+  ///
+
+  // xxx : rename
+  #[ allow( unused_macros ) ]
+  #[ macro_export ]
+  macro_rules! impl_x2_deref
   {
 
-    impl< Scalar > Deref for X2< Scalar >
-    where
-      Scalar : ScalarInterface,
-    {
-      type Target = $Va $( :: $Vb )*< Scalar >;
-      fn deref( &self ) -> &Self::Target
-      {
-        self.as_foreign()
-      }
-    }
+    () => {};
 
-    impl< Scalar > DerefMut for X2< Scalar >
-    where
-      Scalar : ScalarInterface,
+    ( $Va : ident $( :: $Vb : ident )* ) =>
     {
-      fn deref_mut( &mut self ) -> &mut Self::Target
-      {
-        self.as_foreign_mut()
-      }
-    }
 
-  };
+      impl< Scalar > Deref for X2< Scalar >
+      where
+        Scalar : ScalarInterface,
+      {
+        type Target = $Va $( :: $Vb )*< Scalar >;
+        fn deref( &self ) -> &Self::Target
+        {
+          self.as_foreign()
+        }
+      }
+
+      impl< Scalar > DerefMut for X2< Scalar >
+      where
+        Scalar : ScalarInterface,
+      {
+        fn deref_mut( &mut self ) -> &mut Self::Target
+        {
+          self.as_foreign_mut()
+        }
+      }
+
+    };
+
+  }
+
+  #[ allow( unused_imports ) ]
+  pub( crate ) use impl_x2_deref;
 
 }
 
-#[ allow( unused_imports ) ]
-pub( crate ) use impl_x2_deref;
+//
+
+crate::mod_interface!
+{
+  exposed( crate ) use impl_x2_deref;
+}
