+++ module/rust/math_adapter/rust/impl/math/vector/x2/slice.rs
@@ -1,20 +1,32 @@
-
-impl< Scalar > X2NominalInterface for &[ Scalar ]
-where
-  Scalar : ScalarInterface,
+/// Internal namespace.
+pub( crate ) mod private
 {
-  type Scalar = Scalar;
+  use crate::*;
 
-  #[ inline ]
-  fn _0( &self ) -> Self::Scalar
+  impl< Scalar > X2NominalInterface for &[ Scalar ]
+  where
+    Scalar : ScalarInterface,
   {
-    self[ 0 ]
+    type Scalar = Scalar;
+
+    #[ inline ]
+    fn _0( &self ) -> Self::Scalar
+    {
+      self[ 0 ]
+    }
+
+    #[ inline ]
+    fn _1( &self ) -> Self::Scalar
+    {
+      self[ 1 ]
+    }
+
   }
 
-  #[ inline ]
-  fn _1( &self ) -> Self::Scalar
-  {
-    self[ 1 ]
-  }
+}
 
+//
+
+crate::mod_interface!
+{
 }
